Use weak reference for storing rootViews in iOS (#1217)
In `RNGestureHandlerManager.m`, `_rootViews` is a `NSMutableSet` which will retain the elements.

```
NSMutableSet<UIView*> *_rootViews;
```

In fact, the element in _rootViews is `RCTRootContentView`.

When a `RCTRootView` is created, the `RCTRootContentView` will be pushed into the `_rootViews`.

In my case, `RCTBridge` is cached after remove a `RCTRootView`.  `RCTRootView` will dealloc, but `RCTRootContentView` will not dealloc.

Then I allocate a new `RCTRootView` which is inited by my cached `RCTBridge`, a new `RCTRootContentView` will be allocated and it will also be retained by `RNGestureHandlerManager`. At present, the `_rootViews` has two `RCTRootContentView` element, the old and the new. 

I think that the `_rootViews` should not strongly retain the element.

In my opinion, the solution is that using `NSHashTable` instead of `NSMutableSet`

```
NSHashTable<UIView *> *_rootViews;
_rootViews = [NSHashTable hashTableWithOptions:NSPointerFunctionsWeakMemory];
```


Co-authored-by: Norcy <chenying09@kuaishou.com>